feat [bfdr]: Add Remaining Place of Birth IJE/FHIR fields. (#36)
This PR adds the following IJE/FHIR fields:

- BPLACE | BirthPhysicalLocation (Place Where Birth Occurred (type of
place or institution)
- CNTY | PlaceOfBirth['countyC'] (County of Birth (code))
- BPLACEC_CNT | MotherPlaceOfBirth['addressCountry'] (Birthplace of
Mother--Country) [Was already mostly implemented]
- BPLACEC_ST_TER | MotherBirthState (State, U.S. Territory or Canadian
Province of Birth (Mother) - code) [Was mostly already implemented]

Includes the addition of an EncounterBirth in RestoreReferences().
Includes a temporary BFDR valuesets generation script for the place of
birth types. This should be discarded once a complete BFDR valuesets
generation script is merged in.