test(bfdr): add getter and setter unit tests to BirthRecord_Should (#64)
The profile urls in our test files for Mother, Child, and Father were incorrect so they were not getting parsed correctly. Updated the profiles to fix many of the null reference issues. There was also an issue with how the FHIRPath was finding the fields like congenital anomalies that use the category to identify the procedure in the bundle since there could be multiple categories. Some of the test files were missing the category field required for parsing so those were added in. There was also an error in MotherHeightEdit field, the extension was getting added to the observation when it needed to be added to the observation.Value

---------

Co-authored-by: MacAdam <smacadam@mitre.org>