refactor(bfdr): constructor clean up (#68)
## Summary 

Constructors: removing comments to birth constructors, adding others.

Removed race/ethnicity fields and create methods and replaced with
getOrCreateObservation helper.

Am getting two types of errors:

1. One is an exception I put in, so it might be that I should just cut
it out. In BirthRecord_constructors,
`GetDemographicCodedContentBundle()` was returning a bundle with no
composittion (a required field) so I added it along with the required
mother or father sections. If there is no mother or father to report on,
it seems like a pointless record, so I have it raise the exception. Is
this pre-emptive validation inappropriate?
2. The other is a null reference exception in the `SubjectId()` method
in fieldsAndCreateMethods. When I switched to using the
getOrCreateObservation method for race/ethnicity, it was no longer
finding a subject ID when trying to `get` an ethnicity field. Trying to
determine if this is something wrong in the record or an implementation
issue.