feat(bfdr): Update Child/DecedentFetus references, Date of Registration (#80)
This PR:
- Implements fixes for Child/DecendentFetus issues. FetalDeathRecords do
not have a Child profile, so "Child" has been renamed to "Subject" at
the NatalityRecord level and acts as a field for both Child and
DecedentFetus, depending on the record type. There are new overriden
RestoreReferences() and InitializeCompositionAndSubject() methods that
handle setting this Subject field accordingly. A similar function is
performed for the Composition as well.
- - There could be a different approach with overriden
"GetSubjectProfile()" and "GetCompositionProfile()" methods which would
remove the need for overriding "RestoreRefeferences()" again and remove
the need for "InitializeCompositionAndSubject()". Let me know if this
could be a better approach, I just wasn't sure if that could end up
spiraling into a million methods as we add more references that need
populating.
- This PR also includes IJE implementation and tests for Date of
Registration, which is identical with the BirthRecord version.