feat(bfdr): implement fetal death cause or condition (#89)
COD18a1 through COD18a7 are Y/N depending on presence of code
COD18a8 is a literal text/blank

I added two class helper methods, YesNo_PresenceToIJE and
YesNo_IJEToPresence, which are similar to PresenceToIJE and
IJEToPresence but do not require a corresponding none-of-the-above field
(these don't exist for cause of death) and are limited to only Y/N (no
Unknown) per the IJE_File_Layouts.

I also added a helper to VitalRecord_util called GetCondition which is a
modification of GetObservation but for condition resources instead.