feat(bfdr): Delivery Place Type (DPLACE), Certification Date (CERTIFIED_XX) (#86)
This PR implements:

- DPLACE (Place of Delivery) [New implementations]
- - The BPLACE and BirthPhysicalLocation fields are very similar to
DPLACE. The key difference is DPLACE uses EncounterMaternity instead of
EncounterBirth. There were a couple options here, what I decided to do
was refactor out common "Set location in encounter" getter/setter/helper
methods to `VitalRecord_util` and moved BirthPhysicalLocation to
NatalityRecord. I considered an approach where the Encounter to use
would be determined via an abstract method, but I thought this made
things a little confusing and would be inconsistent. It also wouldn't
have allowed for unique property names: there seems to be a strong
distinction between a "Delivery" and a "Birth" in the IG.
- CERTIFIED_ YR/MO/DY (Certification Date) [Copied implementation from
BirthRecord]