ci(bfdr): add code coverage (#138)
## Summary

- add documentation for full code coverage report generation (local) 
- add GH workflow for coverage overview in pull requests


In general we want free, open source tools that confirm test coverage
when creating PRs and give us the ability to look deeper at coverage
locally. The
[CodeCoverageSummary](https://github.com/irongut/CodeCoverageSummary)
generated in this PR was fairly straightforward to set up and is
configurable. This means we can set a threshold for coverage to meet
without the causing the action to fail (and whether or not it can
trigger a failure at all).

For looking closer at the coverage itself, it is useful to have a
user-friendly html render of coverage information. [Coverlet
](https://github.com/coverlet-coverage/coverlet)first generates an xml
file that contains relevant coverage information, and
[ReportGenerator](https://github.com/danielpalme/ReportGenerator?tab=readme-ov-file)
then converts it into navigable html. The report includes cyclomatic
complexity, [CRAP
score](https://testing.googleblog.com/2011/02/this-code-is-crap.html),
line coverage and branch coverage. You can easily see which lines are
coverable, covered, and uncovered across tested files. Coverlet seems to
be widely used ([410.3M
downloads](https://www.nuget.org/packages/coverlet.collector/3.1.1) on
NuGet), well
[documented](https://github.com/coverlet-coverage/coverlet/tree/master/Documentation),
and can be configured as a global tool. ReportGenerator is flexible and
can be used to generate reports from various formats, including for our
use case.

There is sometimes a slight discrepancy between what coverlet and the GH
action coverage summary find, but it looks to be small.

See `readme` for information on running the full code coverage report
locally. Rendered HTML should look like this:

![Screenshot 2024-06-30 at 9 39
33 PM](https://github.com/nightingaleproject/vital-records-dotnet/assets/35311744/e568005c-9e34-42d3-ae84-c634d8b734f2)