638 messaging fetal death (#172)
This PR refactors the existing Messaging classes in BFDR and adds in
fetal death messaging classes.

To cut down on duplicate code, the BaseMessage, Acknowledgement, Status,
Parental Demographic, Error, and Void messages were all renamed from
BirthXMessage to BFDRXMessage to be inclusive of fetal death. Now each
of these message types can be created from a BirthRecordMessage or a
FetalDeathRecord Message.

To support the change above, I requested the Messaging IG be updated to
use "event_year" instead of "birth_year" and "delivery_year" in the
message headers for Birth and FetalDeath messages. The Messaging IG has
been updated with this change.

The GetYear function at the record level was updated to the Messages can
pull the year from the underlying record.