fix(vr): 722 remove use of meta in the library and 710 fix mother birthplace in canary (#221)
This pull request addresses NVSS-722: meta profile tags are not required
to be present in FHIR documents, so should not be relied on when reading
content from a VR record of any type; these changes update the code to
use alternate approaches for finding content such as following
references.

This pull request also provides a work around for NVSS-710. Canary uses
FHIR paths to identify and display relevant sections of a FHIR document.
The 4.3.0 version of the Hl7.Fhir.R4 library does not support following
references in FHIR paths; until a newer version is available we make an
exception to our avoidance of relying on meta profile tags since there
is no other way to distinguish between the patient entries for the
mother and the child.

Specific changes include

1. Updating fixtures to use correct type codes and include required
content like compositions
2. Updating BFDR constructors to use correct type codes for compositions
3. Updating FHIR paths to avoid the use of meta.profile where possible
(besides the one exception mentioned above)
4. Updating BFDR RestoreReferences() to reflect that BFDR records always
have a composition, determine whether a subject is expected based on the
composition type, and use references to find content rather than meta
profiles
5. Updating VRDR RestoreReferences() to use the bundle type to determine
whether a full record is present