fix(bfdr): update fakerecord fetaldeath to include more data (#229)
This PR addresses ticket
[NVSS-743](https://tracker.codev.mitre.org/browse/NVSS-743)

To properly test fetal death testing with NCHS, we need to provide
quality data examples with enough data for NCHS to create coded content
for cause of death and demographic information. Veronique provided a
list of fields to make sure are included in the test data. This PR
updates the fakerecord CLI tool to add the following data to the
fakerecord so we can generate quality test examples:

- Sex  
- Race and ethnicity
- Cause or condition
- Place where delivery occurs 
- Date of Birth for Mother are mapped
- Mother Residence address

The ticket also included reviewing our fetal death test fixtures and
updating them to improve the data quality. Each test fixture was updated
to
- include race and ethnicity data for the mother and father
- include sex
- include a cause or condition
- include the place of delivery
- include mother birthdate
- include mother residence

A lot of the submission examples or records included coded fields. These
were probably added during development, but don't make sense since coded
fields aren't part of fetal death record submissions. I removed the
coded sections from the fetal death record fixtures where appropriate.


UPDATE 4/11: 
These changes uncovered a flaw in how we create the sections of the
composition and set the focus id to the Mother Patient ID. We rely on
the focus id to parse the Mother Patient. This PR updates the natality
constructors to initialize the composition sections, set the focus id,
and provide a default empty reason. The empty reason is removed when an
entry is added to the composition section.