fix(bfdr): ensure numeric fields are padded with leading zeroes (#242)
Many numeric IJE fields require leading zeros to pad the field to the
required length. E.g. the HIN field is two characters and a length of
four should be returned as "04" not "4". This PR fixes the return value
for IJE field HIN and HFT in both birth and fetal death records.

Fixes #NVSS-781