fix(bfdr): deleted fields return blank strings
Fields in IJE types that are marked as deleted by NCHS return blank
strings of the normal field length when accessed. Remove the
corresponding fields from the FHIR birth and fetal death record types.
Fields that are now exclusive to birth records are owned by the FHIR
birth record type.