fix: change how combinations of Y/N/U flags work
IJE fields represented with Y, N, and U can represent attributes that
are part of the same category in FHIR. However, not all combinations of
these values can be represented in FHIR. When setting Y for an IJE
field, add the corresponding FHIR resource, and remove the "none of the
above" entry if it is present. When setting N for an IJE field, remove
the corresponding FHIR resource. If that field's category is now empty,
add the corresponding "none of the above" entry. When setting U for an
IJE field, remove the corresponding FHIR resource, but do not check the
field's category. The order in which the fields are processed does not
affect the resulting records.

Revert previous changes to connectathon IJE records.

Add tests for IJEBirth to demonstrate the expected behavior for
different combinations of Y, N, and U.