feat(canary): show errors for text area inputs below the text area (#252)
The Getter component is used to provide a text area and display a list
of issues. Display any error returned by the API server using the list
of issues.

The original goal of this task was to do a more complete overhaul of
error messages displayed in Canary. But, as I learned more about the
variety of errors and API routes, this seemed less feasible. So for this
PR, I reduced the scope to the `Getter` component. Other considerations
for errors shown in Canary include:
* Errors returned when using the Connectathon component, as various
parts of the interface are hidden/shown throughout the usage of the
component.
* Uncaught exceptions from API routes. Some exceptions are from data
errors that the user can fix, but other errors indicate problems that
the user cannot fix (and the user should probably report them to
someone)
* Some API handling functions are designed to only return a Record,
while others return a Record and a list of issues.

Addresses Jira issue 761.