fix: change how combinations of Y, N, and U work in IJE records (#249)
~Change IJE fields for flags in the same category so that combinations
of Y and N are replaced by combinations of Y and U, since N should only
be used when none of the flags are set to Y.~ This solution was
incorrect and has been replaced in commit
2dbb17f8119402268bf93aaa9881aa0055c722d2.

IJE fields represented with Y, N, and U can represent attributes that
are part of the same category in FHIR. However, not all combinations of
these values can be represented in FHIR. When setting Y for an IJE
field, add the corresponding FHIR resource, and remove the "none of the
above" entry if it is present. When setting N for an IJE field, remove
the corresponding FHIR resource. If that field's category is now empty,
add the corresponding "none of the above" entry. When setting U for an
IJE field, remove the corresponding FHIR resource, but do not check the
field's category. The order in which the fields are processed does not
affect the resulting records.

Add tests for IJEBirth to demonstrate the expected behavior for
different combinations of Y, N, and U.

Addresses Jira issue 793.