chore: add windows versions of run_tests scripts (#255)
The `run_tests.bat` scripts run all the same commands as the existing
`run_tests.sh` scripts. There's no way to specify "exit when a command
returns an error code" for an entire file in a batch script, so each
command has a suffix that causes it to exit upon failure.

Addresses Jira issue 774.